+++ debezium-server-cratedb/src/main/java/io/debezium/server/cratedb/ColumnName.java
@@ -8,6 +8,8 @@ package io.debezium.server.cratedb;
 import java.util.List;
 import java.util.Objects;
 
+import io.debezium.server.cratedb.types.ColumnType;
+
 public record ColumnName(String columnName) {
     public ColumnName {
         Objects.requireNonNull(columnName);
