+++ debezium-server-cratedb/src/main/java/io/debezium/server/cratedb/types/ArrayType.java
@@ -3,7 +3,7 @@
  *
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package io.debezium.server.cratedb;
+package io.debezium.server.cratedb.types;
 
 public record ArrayType(ColumnType elementType) implements ColumnType {
     @Override
