+++ debezium-server-cratedb/src/main/java/io/debezium/server/cratedb/types/BooleanType.java
@@ -3,12 +3,12 @@
  *
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package io.debezium.server.cratedb;
+package io.debezium.server.cratedb.types;
 
-public record BigIntType() implements ColumnType {
+public record BooleanType() implements ColumnType {
     @Override
     public String shortName() {
-        return "bigint";
+        return "boolean";
     }
 
     @Override
